refactor(widgets): rename of files
