+++ src/widgets/header/ui/close-menu.tsx
